fixing rank in infinite scroll
