+++ src/features/colinks/useInfiniteMembers.tsx
@@ -1,3 +1,5 @@
+import { useCallback, useEffect } from 'react';
+
 import { QueryKey, useInfiniteQuery } from 'react-query';
 
 import { ValueTypes } from '../../lib/gql/__generated__/zeus';
@@ -41,12 +43,13 @@ const getMembers = async (
 };
 
 export const useInfiniteMembers = (
+  observerRef: React.MutableRefObject<HTMLDivElement | null>,
   currentAddress: string,
   queryKey: QueryKey,
   where: Where,
   orderBy: OrderBy[]
 ) => {
-  return useInfiniteQuery(
+  const iq = useInfiniteQuery(
     queryKey,
     ({ pageParam = 0 }) =>
       getMembers(currentAddress, where, orderBy, pageParam),
@@ -58,4 +61,31 @@ export const useInfiniteMembers = (
       refetchInterval: 10000,
     }
   );
+
+  const { fetchNextPage, hasNextPage } = iq;
+
+  const handleObserver = useCallback<
+    (entries: IntersectionObserverEntry[]) => void
+  >(
+    entries => {
+      const [target] = entries;
+      if (target.isIntersecting) {
+        fetchNextPage();
+      }
+    },
+    [fetchNextPage, hasNextPage]
+  );
+
+  useEffect(() => {
+    const element = observerRef.current;
+    if (element) {
+      const option = { threshold: 0 };
+
+      const observer = new IntersectionObserver(handleObserver, option);
+      observer.observe(element);
+      return () => observer.unobserve(element);
+    }
+  }, [fetchNextPage, hasNextPage, handleObserver]);
+
+  return { ...iq };
 };
